Starting project-wide cleanup pass after hiatus.
Back on this project after a hiatus.  I'm going to do a full pass over it;  trying to clean it up and implement better architecture.  The goal of version 1 was to simply "get it working."  That was accomplished, so now the goal of version 1.1 will be to make it a better "example project."  That means clean code and a straightforward architecture.  Compared to "proper" architecture, shortcuts will still be taken in the name of simplicity, which are reasonable considering the extremely small scope of the project.

In this first set of changes, I've moved important app classes (assets, camera) into public static fields for ease of use.   In order to make the individual screens simpler, shared code was pulled into an abstract base class.  The custom Asset Manager (non lib-gdx) was removed since it really didn't add anything.  Many files received small updates for organization/cleanliness.